requirements.txt
# GUI Framework
streamlit==1.51.0
streamlit-autorefresh==1.0.1
streamlit-searchbox==0.1.23
streamlit-keyup==0.3.0
pillow==10.3.0

# Database
sqlalchemy==2.0.30

# Google Vision API for OCR and Image Search
google-cloud-vision==3.7.2

# Image Processing
opencv-python==4.9.0.80
numpy==1.26.4

# Invoice Generation
reportlab==4.2.0
openpyxl==3.1.2

# PDF Processing
PyPDF2==3.0.1
pdf2image==1.17.0

# Excel/CSV Processing
pandas==2.2.2

# Utilities
python-dotenv==1.0.1
requests==2.31.0

# Image Similarity Search (for AI search)
scikit-learn==1.4.2
sentence-transformers==2.7.0

# Date/Time
python-dateutil==2.9.0.post0
